refactor: hive_initializer 이동
